pratice7_11

+++ usth book/Avg_num.java
@@ -15,12 +15,12 @@ public class Avg_num{
         num.add(In_num);
     }
     public void avg(int n){
-        Integer sum = 0;
+        int sum = 0;
         for(Integer i:num){
             sum += i;
             System.out.printf("%d ", i);
         }
-        Integer avg_n = sum/n;
+        double avg_n = sum/n;
         System.out.println("average of number is:" +avg_n);
     }
